+++ src/commands/mod.rs
@@ -9,4 +9,5 @@ pub mod get_totals;
 pub mod info;
 pub mod list_ports;
 pub mod set_home_altitude;
+pub mod set_sleep_screen;
 pub mod update;